requirements.txt
pandas==2.1.4
coloredlogs~=15.0.1
mysql-connector-python~=8.2.0